tests/wpt/tests/credential-management/fedcm-revoke.https.html

<!DOCTYPE html>
<title>Federated Credential Management API revoke() tests.</title>
<link rel="help" href="https://fedidcg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<body>

<script type="module">
import {fedcm_test,
        revoke_options,
        fedcm_get_and_select_first_account,
        request_options_with_mediation_required} from './support/fedcm-helper.sub.js';

fedcm_test(async t => {
  const revoke = IdentityCredential.revoke(revoke_options("nonExistent"));
  return promise_rejects_dom(t, 'NetworkError', revoke);
}, 'Test that revoke fails when there is no account to revoke');

fedcm_test(async t => {
  const cred = await fedcm_get_and_select_first_account(t, request_options_with_mediation_required());

  return IdentityCredential.revoke(revoke_options("1234"));
}, 'Test that revoke succeeds when there is an account to revoke');

fedcm_test(async t => {
  const cred = await fedcm_get_and_select_first_account(t, request_options_with_mediation_required());

  await IdentityCredential.revoke(revoke_options("1234"));

  const revoke = IdentityCredential.revoke(revoke_options("1234"));
  return promise_rejects_dom(t, 'NetworkError', revoke);
}, 'Test that revoking the same account twice results in failure.');

</script>
